ReactWebsite/syllabus-main/requirements.txt
fitz
pandas
numpy
matplotlib.pyplot
os
sys
load_dotenv
OpenAI
